Add files via upload
The following project consists of the implementation of a Spear Phishing attack scenario in a controlled environment using virtual machines: the attacker was impersonated by a Kali Linux virtual machine, configured with a network card in Bridge mode, hosted by a Windows 11 operating system.
The victim was represented by a Windows 7 operating system. Specifically, the target of our attack was a reverse shell on the victim machine, with the intent to take over. The attack payload was produced by means of Metasploit and appropriately coded, while the attack vector was an Excel document where VBA code( Excel macros) was inserted : a bogus email with the malicious document attached is delivered to the victim who by downloading the document, unintentionally, allows us access to his laptop.